# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Apple Store API
  description: API Document for the Applestore project based on the ERD.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Users
    description: User management
  - name: Categories
    description: Category management
  - name: Products
    description: Product management
  - name: Orders
    description: Order management
  - name: OrderItems
    description: Order item management

paths:
  # ==========================================
  # USERS API (CRUD)
  # ==========================================
  /api/v1/users:
    get:
      tags: [Users]
      summary: Get all users
      operationId: getAllUsers
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseUserDTO'
    post:
      tags: [Users]
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUserDTO'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUserDTO'

  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUserDTO'
    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUserDTO'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUserDTO'
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully (No Content)

  # ==========================================
  # CATEGORIES API (CRUD)
  # ==========================================
  /api/v1/categories:
    get:
      tags: [Categories]
      summary: Get all categories
      operationId: getAllCategories
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseCategoryDTO'
    post:
      tags: [Categories]
      summary: Create a new category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCategoryDTO'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCategoryDTO'

  /api/v1/categories/{id}:
    get:
      tags: [Categories]
      summary: Get category by ID
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: Category ID as a string (e.g., CAT-GIFT-01)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCategoryDTO'
    put:
      tags: [Categories]
      summary: Update category
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCategoryDTO'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseCategoryDTO'
    delete:
      tags: [Categories]
      summary: Delete category
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully (No Content)

  # ==========================================
  # PRODUCTS API (CRUD)
  # ==========================================
  /api/v1/products:
    get:
      tags: [Products]
      summary: Get all products
      operationId: getAllProducts
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseProductDTO'
    post:
      tags: [Products]
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestProductDTO'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseProductDTO'

  /api/v1/products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: The product ID (e.g., PROD-BOX-001)
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseProductDTO'
    put:
      tags: [Products]
      summary: Update product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestProductDTO'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseProductDTO'
    delete:
      tags: [Products]
      summary: Delete product
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully (No Content)

  # ==========================================
  # ORDERS API (CRUD)
  # ==========================================
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: Get all orders
      operationId: getAllOrders
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseOrderDTO'
    post:
      tags: [Orders]
      summary: Place a new order (Create Order with OrderItems)
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOrderDTO'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseOrderDTO'

  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details by ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseOrderDTO'
    put:
      tags: [Orders]
      summary: Update order status
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOrderDTO'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseOrderDTO'
    delete:
      tags: [Orders]
      summary: Delete or cancel order
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully (No Content)

  # ==========================================
  # ORDER ITEMS API (CRUD)
  # ==========================================
  /api/v1/order-items:
    get:
      tags: [OrderItems]
      summary: Get all order items
      operationId: getAllOrderItems
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseOrderItemDTO'
    post:
      tags: [OrderItems]
      summary: Create a new order item
      operationId: createOrderItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCreateOrderItemDTO'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseOrderItemDTO'

  /api/v1/order-items/{id}:
    get:
      tags: [OrderItems]
      summary: Get order item by ID
      operationId: getOrderItemById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseOrderItemDTO'
    put:
      tags: [OrderItems]
      summary: Update order item
      operationId: updateOrderItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdateOrderItemDTO'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseOrderItemDTO'
    delete:
      tags: [OrderItems]
      summary: Delete order item
      operationId: deleteOrderItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully (No Content)

# ==========================================
# SCHEMAS (DTOs)
# ==========================================
components:
  schemas:
    # --- USER DTOs ---
    RequestUserDTO:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string
      required: [email, password, first_name, last_name]

    ResponseUserDTO:
      type: object
      description: Note - Never return the password to the client.
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string

    # --- CATEGORY DTOs ---
    RequestCategoryDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]

    ResponseCategoryDTO:
      type: object
      properties:
        id:
          type: string
          description: Maps to category_id (e.g., CAT-GIFT-01)
        name:
          type: string
        description:
          type: string

    # --- FLAVOR PROFILE DTO (Nested in Product) ---
    FlavorProfileDTO:
      type: object
      properties:
        sweetness_level:
          type: integer
          format: int32
        tartness_level:
          type: integer
          format: int32
        overall_profile:
          type: string
        dominant_notes:
          type: array
          items:
            type: string
        tasting_description:
          type: string

    # --- PRODUCT DTOs ---
    RequestProductDTO:
      type: object
      properties:
        category_id:
          type: string
        name:
          type: string
        description:
          type: string
        tasting_notes:
          type: array
          items:
            type: string
        best_for:
          type: array
          items:
            type: string
        origin:
          type: string
        season:
          type: string
        flavor_profile:
          $ref: '#/components/schemas/FlavorProfileDTO'
        price:
          type: number
          format: double
        stock_quantity:
          type: integer
          format: int32
        image_url:
          type: string
      required: [category_id, name, price, stock_quantity]

    ResponseProductDTO:
      type: object
      properties:
        id:
          type: string
          description: Maps to product_id (e.g., PROD-BOX-001)
        category:
          $ref: '#/components/schemas/ResponseCategoryDTO'
        name:
          type: string
        description:
          type: string
        tasting_notes:
          type: array
          items:
            type: string
        best_for:
          type: array
          items:
            type: string
        origin:
          type: string
        season:
          type: string
        flavor_profile:
          $ref: '#/components/schemas/FlavorProfileDTO'
        price:
          type: number
          format: double
        stock_quantity:
          type: integer
          format: int32
        image_url:
          type: string
        created_at:
          type: string
          format: date-time

    # --- ORDER ITEM DTOs ---
    RequestOrderItemDTO:
      type: object
      description: When placing an order, the frontend only needs to send the product ID and quantity.
      properties:
        product_id:
          type: string
        quantity:
          type: integer
          format: int32
      required: [product_id, quantity]

    RequestCreateOrderItemDTO:
      type: object
      properties:
        order_id:
          type: integer
          format: int64
        product_id:
          type: string
        quantity:
          type: integer
          format: int32
      required: [order_id, product_id, quantity]

    RequestUpdateOrderItemDTO:
      type: object
      properties:
        product_id:
          type: string
        quantity:
          type: integer
          format: int32
      required: [product_id, quantity]

    ResponseOrderItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product_id:
          type: string
        quantity:
          type: integer
          format: int32
        unit_price:
          type: number
          format: double

    # --- ORDER DTOs ---
    RequestOrderDTO:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
          description: This field can be omitted if the backend extracts user_id directly from the JWT Token.
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/RequestOrderItemDTO'
      required: [user_id, order_items]

    ResponseOrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        order_date:
          type: string
          format: date-time
        total_amount:
          type: number
          format: double
        status:
          type: string
          enum: [PENDING, SHIPPED, DELIVERED]
        order_items:
          type: array
          items:
            $ref: '#/components/schemas/ResponseOrderItemDTO'